--- src/w250210/예진/p16236/Main.java
@@ -1,4 +1,0 @@
-package w250210.예진.p16236;
-
-public class Main {
-}
